+++ CHANGELOG.md
@@ -10,8 +10,10 @@
 - MapMarker: Added the ability to additionally highlight polygons
 - FullScreen: added "newTabFromFrame" configuration, to disable opening a new tab if in an iFrame
 - LayerTree: added "contactPublisherName" configuration, to change the info message shown, when clicking the contact sub menu.
+- UrlParams: Added URL-Params mainclosed and secondaryclosed to have the menus initially closed when set to true and opened when set to false.
 
 ### Changed
+- WMS-Time: Shrunk wms-time slider for mobile views, adjusted Position on desktop.
 - getFeatureInfo: gfi now works on BaseLayers as well as SubjectLayers.
 
 ### Deprecated